refactored  ridge blockout viewer into multiple files
